+++ utils/calculate_step.c
@@ -1,6 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   calculate_step.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ribana-b <ribana-b@student.42malaga.com>   +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/18 21:26:28 by ribana-b          #+#    #+# Malaga      */
+/*   Updated: 2025/01/18 21:26:29 by ribana-b         ###   ########.com      */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "utils.h"
 
-int	calculate_step(float dx, float dy)
+int	calculate_step(double dx, double dy)
 {
 	if (dx < 0)
 		dx *= -1;
