+++ utils/v2/v2div.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   v2div.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ribana-b <ribana-b@student.42malaga.com>   +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/18 21:25:04 by ribana-b          #+#    #+# Malaga      */
+/*   Updated: 2025/01/18 21:25:16 by ribana-b         ###   ########.com      */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "utils.h"
 
 t_v2	v2_div(t_v2 *v1, t_v2 *v2)
@@ -8,7 +20,7 @@ t_v2	v2_div(t_v2 *v1, t_v2 *v2)
 	});
 }
 
-t_v2	v2_div_scalar(t_v2 *v1, float scalar)
+t_v2	v2_div_scalar(t_v2 *v1, double scalar)
 {
 	return ((t_v2){
 		.x = v1->x / scalar,
